tests: xml: add realistic XML tests files
This patch refresh current XML testfiles with some realworld
expressions extracted from rules. The nft instruction itself is added
as a comment for future references.

All XMl files are now indented with tabs instead of spaces. Also, a
bunch of new realworld rules with mixed expressions are added.

I used this command to get the XML formatted with tabs:

$ export XMLLINT_INDENT=$'\t'
$ xmllint --format file.xml

The xmllint tool is included in the libxml2-utils package (at least on
debian systems).

Signed-off-by: Arturo Borrero Gonzalez <arturo.borrero.glez@gmail.com>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>