src: remove version from XML and JSON representations
This patch removes the version XML node and the version JSON field
in all our existing objects. The current versioning approach
consists of adding a version field to every object representation
in XML and JSON. While listing my entire rule-set, one can notice
that this approach is too bloated.

Once the library enters stable stage, if we need to obsolete a XML
node and a JSON field, we can follow this procedure:

1) Remove the XML node and the JSON field from the output, so fresh
   outputs will not contain the old ones anymore.

2) Do not remove the parsing of the old XML node and the JSON field
   inmediately. We have to keep supporting the parsing for a while
   to avoid breaking the interpretion of old XML/JSON files. We can
   spot a warning to warn about it, so users generate a fresh
   output again.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>