tests: xml: remove \t and \n from all the files
These XML files are intented to be used for automated tests. Thus, it's
important to have exactly the same output that was generated by the library.

The beautified output does not match the real output that the library provides
and it just makes things harder for us to provide a more advanced automated
testing.

If you want to make any of those file human readable, please use:

    xmllint --format file.xml

or open the xml file and any browser, eg. firefox.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>