tests: better parsing validation
Before this patch, the test only verified that we could parse
the file. After this, it also compares what it has parsed with
the original file.

Basically, the idea is that:

file -> buffer -> object -> buffer'
   read               snprintf

So we also compare that buffer = buffer' which should be true
since our tests should be based on the output that the library
generates.

This is an example output for the OK and FAILED cases:

parsing and validating xmlfiles/25-rule-exthdr.xml: OK
validating xmlfiles/27-rule-limit.xml: FAILED
from file:     it"><rate>123123</ra
from snprintf: it"><rate>321321</ra

Signed-off-by: Alvaro Neira Ayuso <alvaroneay@gmail.com>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>