expr: log: Do not print unset values in json
It changes the parse and the snprint functions to omit unset values.

Also, It fixes an unnecessary comma after key-value pair type.
This comma is not necessary if there is not more key-value pairs in this expr.

Example:
"expr":[{"type":"log"}]

If It uses this rule:
nft add rule ip test output log

It gets this json file:

[...]
{"expr":[{"type":"log","prefix":"(null)","group":0,"snaplen":0,"qthreshold":0}]}
[...]

Now, That rule creates this json file without null values:

{"expr":[{"type":"log"}]}

Signed-off-by: Ana Rey <anarey@gmail.com>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>