src: add support to import JSON/XML with the new command tag
This patch adds support to parse the new command tag in XML/JSON. This
patch adds two new functions:

* nft_ruleset_parse_file_cb
* nft_ruleset_parse_buffer_cb

The idea is to invoke the callback function that is passed as parameter is
called for each object that is parsed from the corresponding input. Each
callback has access to the nft_parse_ctx structure that provides the necessary
context such as the command, the object type and the object itself.

This change also adds support to update the content of a set incrementally.

	{"nftables":[{"add":[{"element":{"name":"blackhole","table":"filter",
	 "family":"ip","key_type":7,"key_len":4,"set_elem":[{"key":{
	 "reg":{"type":"value","len":4,"data0":"0x0403a8c0"}}}]}}]}]}

This also patch consolidates the xml/json ruleset import path.

Signed-off-by: Alvaro Neira Ayuso <alvaroneay@gmail.com>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>